chore(Database) - Docker-Compose Support of Database
Brings in support for running the database in docker.

The sensitive data is stored in the .gitignore'd file.

To run, simply navigate using your Docker host
(in this case WSL2 with Docker Desktop for Windows installed)
and then use the following commands:

       cd ......../DigitalInspection/db-docker
       docker-compose up

To turn off, use

       docker-compose down

To re-seed DB, add on or replace DB files in docker-compose.yml.
It is applied alphabetically.